refactor chatwitheino event helpers
